deleted authform and renamed AuthFormSimple
